==> Scripts/04-01-Requetes_SQL_Bdd.ipynb <==
# Librairie Os pour les chemins de fichiers et variables d'environnement
import os
# Librairie csv pour créer et lire le format csv
import csv
# Lien au fichier Python pour exécuter ses fonctions
import functions_PgSql
# import Donnees
import psycopg2
import pandas as pd
import numpy as np

 
# Spécifier le nom de la base de données créée au préalable dans PgAdmin
ma_base_donnees = "Musees_V3"
utilisateur = "postgres"

# Méthode os.environ.get pour utiliser le mot de passe 
# enregistré au préalable dans une variable d'environnement
mot_passe = os.environ.get('pg_psw')

# Fonction conn appelle la fonction ouvrir_connexion du fichier annexe Python
conn = functions_PgSql.ouvrir_connection(ma_base_donnees, utilisateur, mot_passe)


## VUES SQL

# Création vue Oeuvres avec infos Artistes et Titres d'oeuvres
creation_vue_Oeuvres_Artistes = """
        CREATE VIEW v_Oeuvres_Artistes (
        Nom_Artiste,
        Id_Oeuvre,
        Titre_Oeuvre,
        Denomination_Oeuvre,
        Materieux_Techniques_Oeuvre,
        Sujet_Represente,
        Periode_Oeuvre,
        Id_Museo
        )
    AS
        SELECT a.nom_artiste, 
               o.id_oeuvre,
               o.titre_oeuvre, 
               o.denomination, 
               o.materiaux_techniques, 
               o.sujet_represente, 
               o.periode_creation, 
               o.id_museo
        FROM oeuvre o, oeuvre_artiste oa, artiste a
        
        WHERE o.id_oeuvre = oa.id_oeuvre
        AND oa.id_artiste = a.id_artiste;
"""

functions_PgSql.executer_requete(creation_vue_Oeuvres_Artistes, conn)

artistes_oeuvres = pd.read_sql_query("SELECT * FROM v_Oeuvres_Artistes WHERE nom_artiste LIKE '%RUBENS%' ORDER BY id_museo;", conn)

artistes_oeuvres.head()

# Création vue Oeuvres avec infos Catégories
creation_vue_Oeuvres_Categories = """
        CREATE VIEW v_Oeuvres_Categories (
        Nom_Categorie,
        Id_oeuvre,
        Titre_Oeuvre,
        Denomination_Oeuvre,
        Materieux_Techniques_Oeuvre,
        Sujet_Represente,
        Periode_Oeuvre,
        Id_Museo
        )
    AS
        SELECT c.domaine, 
               o.id_oeuvre,
               o.titre_oeuvre, 
               o.denomination, 
               o.materiaux_techniques, 
               o.sujet_represente, 
               o.periode_creation, 
               o.id_museo
        FROM oeuvre o, oeuvre_categorie oc, categorie c
        
        WHERE o.id_oeuvre = oc.id_oeuvre
        AND oc.id_categorie = c.id_categorie;
"""

functions_PgSql.executer_requete(creation_vue_Oeuvres_Categories, conn)

oeuvres_categories = pd.read_sql_query("SELECT * FROM v_Oeuvres_Categories WHERE Nom_Categorie LIKE '%%';", conn)

oeuvres_categories.head()

# Infos artistes et titres des oeuvres
# Join table de relation many to many oeuvre_artiste
requete = '''
    SELECT o.titre_oeuvre, a.*
    FROM oeuvre o
    
    RIGHT JOIN oeuvre_artiste oa
        ON o.id_oeuvre = oa.id_oeuvre 
    RIGHT JOIN artiste a
        ON oa.id_artiste = a.id_artiste
    
    ORDER BY titre_oeuvre;
'''
Info_Artiste_Titre = pd.read_sql_query(requete, conn)
print(Info_Artiste_Titre.shape)
Info_Artiste_Titre.head()

# Titre et nom artiste
requete = '''
    SELECT titre_oeuvre, nom_artiste
    FROM oeuvre o
    
    JOIN oeuvre_artiste oa
        ON o.id_oeuvre = oa.id_oeuvre 
    JOIN artiste a
        ON oa.id_artiste = a.id_artiste
    
    ORDER BY titre_oeuvre;
'''
Titre_Artiste = pd.read_sql_query(requete, conn)
print(Titre_Artiste.shape)
Titre_Artiste.head()

# Nombre d'oeuvres réalisées par artistes
# Recherche par nom
requete = '''
    SELECT DISTINCT nom_artiste, 
           COUNT(o.id_oeuvre) AS Nb_Oeuvre
    FROM oeuvre o
    
    JOIN oeuvre_artiste oa
        ON o.id_oeuvre = oa.id_oeuvre 
    JOIN artiste a
        ON oa.id_artiste = a.id_artiste
    
    WHERE nom_artiste <> 'ANONYME'
    
    GROUP BY nom_artiste
    
    HAVING COUNT(o.id_oeuvre) >= 100
    
    ORDER BY nb_oeuvre DESC;
'''
Nb_Oeuvres_Artiste = pd.read_sql_query(requete, conn)
print(Nb_Oeuvres_Artiste.shape)
Nb_Oeuvres_Artiste.head()

# Quels Artistes exposés par Ville
# Recherche par Ville, Infos Musées et noms artistes
# Utilisation de la vue v_Oeuvres_Artistes avec une autre jointure
requete = """
    SELECT DISTINCT a.Nom_Artiste,
           COUNT(a.id_oeuvre) AS Nb_Oeuvres,
           m.nom_officiel AS Nom_Musee,
           m.ville
    FROM v_Oeuvres_Artistes a
    
    JOIN musee m 
        ON a.Id_Museo = m.Id_Museo
    
    WHERE m.ville = 'Nantes'
    GROUP BY a.Nom_Artiste, m.nom_officiel, m.ville
    ORDER BY Nom_Musee, Nb_Oeuvres DESC;
"""
Localiser_Ville = pd.read_sql_query(requete, conn)
print(Localiser_Ville.shape)
Localiser_Ville.head()

# Quels Musées Exposent un Artiste
# Recherche par Nom Artiste, Infos Musées et noms artistes
# Utilisation de la vue v_Oeuvres_Artistes avec une autre jointure
requete = """
    SELECT DISTINCT a.Nom_Artiste,
           m.nom_officiel AS Nom_Musee,
           m.ville
    FROM v_Oeuvres_Artistes a
    
    JOIN musee m 
        ON a.Id_Museo = m.Id_Museo
    
    WHERE  a.Nom_Artiste LIKE '%RENOI%'
    
    ORDER BY Nom_Musee, ville;
"""
Localiser_Artiste = pd.read_sql_query(requete, conn)
print(Localiser_Artiste.shape)
Localiser_Artiste.head()

# Quels Musées Exposent un Artiste et nombre d'oeuvres par Artistes
# Recherche par Nom Artiste
# Utilisation de la vue v_Oeuvres_Artistes avec une autre jointure
requete = """
    SELECT a.Nom_Artiste, 
           COUNT(DISTINCT a.id_oeuvre) AS Nb_Oeuvres,
           m.nom_officiel AS Nom_Musee,
           m.ville
    FROM v_Oeuvres_Artistes a
    
    JOIN musee m 
        ON a.Id_Museo = m.Id_Museo
    
    WHERE  a.Nom_Artiste LIKE '%VINCI%'
    
    GROUP BY a.Nom_Artiste, m.nom_officiel, m.ville 
    
    ORDER BY nb_oeuvres DESC;
"""
Localiser_Artiste_Nb_Oeuvres = pd.read_sql_query(requete, conn)
print(Localiser_Artiste_Nb_Oeuvres.shape)
Localiser_Artiste_Nb_Oeuvres.head(38)

# Quels Musées Exposent un Artiste sur une période donnée
# Recherche par période : siècle, Infos Musées et noms artistes et nombre d'oeuvres exposées
# Utilisation de la vue v_Oeuvres_Artistes avec une autre jointure
requete = """
    SELECT a.Nom_Artiste, 
           COUNT(a.id_oeuvre) AS Nb_Oeuvres,
           a.periode_oeuvre,
           m.nom_officiel AS Nom_Musee,
           m.ville
    FROM v_Oeuvres_Artistes a
    
    JOIN musee m 
        ON a.Id_Museo = m.Id_Museo
    
    WHERE  a.Periode_Oeuvre LIKE '%21%'
    
    GROUP BY a.Nom_Artiste, 
             a.periode_oeuvre, 
             m.nom_officiel, 
             m.ville
    
    ORDER BY ville;
"""
Localiser_Periode = pd.read_sql_query(requete, conn)
print(Localiser_Periode.shape)
Localiser_Periode.head()

# Poucentage d'oeuvres exposées par villes / total oeuvres
requete = """
    SELECT m.ville,
           COUNT(o.id_oeuvre) AS Total_Oeuvres,
           ROUND(ROUND(COUNT(o.id_oeuvre),2)*100/(SELECT ROUND(COUNT(o.id_oeuvre),2) FROM oeuvre o),2) || ' %' AS Part_Oeuvres,
           COUNT(DISTINCT o.Id_Museo) AS Nb_Musees
    FROM oeuvre o,
    	musee m
 	
    WHERE o.Id_Museo = m.Id_Museo
    
    GROUP BY ville
    ORDER BY total_oeuvres DESC;
"""
Part_Oeuvres = pd.read_sql_query(requete, conn)
print(Part_Oeuvres.shape)
Part_Oeuvres.head()

# Poucentage d'oeuvres exposées par région, total oeuvres et musées
requete = """
    SELECT r.nom_region,
           COUNT(DISTINCT o.id_museo) AS Nb_Musees,
           COUNT(o.id_oeuvre) AS Total_Oeuvres,
           ROUND(ROUND(COUNT(o.id_oeuvre),2)*100/(SELECT ROUND(COUNT(o.id_oeuvre),2) FROM oeuvre o),2) || ' %' AS Part_Oeuvres
    FROM oeuvre o,
    	 musee m,
         departement d,
         region r
 	
    WHERE o.Id_Museo = m.Id_Museo
    AND m.code_dpt = d.id_dpt
    AND d.id_region = r.id_region
    
    GROUP BY nom_region
    ORDER BY total_oeuvres DESC;
"""
Région_Oeuvres = pd.read_sql_query(requete, conn)
print(Région_Oeuvres.shape)
Région_Oeuvres.head(30)

# Poucentage d'oeuvres exposées par départements, total oeuvres et musées
requete = """
    SELECT d.nom_departement,
           COUNT(DISTINCT o.id_museo) AS Nb_Musees,
           COUNT(o.id_oeuvre) AS Total_Oeuvres,
           ROUND(ROUND(COUNT(o.id_oeuvre),2)*100/(SELECT ROUND(COUNT(o.id_oeuvre),2) 
                                                  FROM oeuvre o),2) || ' %' AS Part_Oeuvres
    FROM oeuvre o,
    	 musee m,
         departement d
 	
    WHERE o.Id_Museo = m.Id_Museo
    AND m.code_dpt = d.id_dpt
    
    GROUP BY nom_departement
    ORDER BY total_oeuvres DESC;
"""
Dpt_Oeuvres = pd.read_sql_query(requete, conn)
print(Dpt_Oeuvres.shape)
Dpt_Oeuvres.head()

# Poucentage d'oeuvres exposées par villes, total oeuvres et musées
requete = """
    SELECT m.ville,
           COUNT(DISTINCT o.id_museo) AS Nb_Musees,
           COUNT(o.id_oeuvre) AS Total_Oeuvres,
           ROUND(ROUND(COUNT(o.id_oeuvre),2)*100/(SELECT ROUND(COUNT(o.id_oeuvre),2) 
                                                  FROM oeuvre o),2) || ' %' AS Part_Oeuvres
    FROM oeuvre o,
    	 musee m
 	
    WHERE o.Id_Museo = m.Id_Museo
        
    GROUP BY ville
    ORDER BY nb_musees DESC;
"""
Villes_Oeuvres = pd.read_sql_query(requete, conn)
print(Villes_Oeuvres.shape)
Villes_Oeuvres.head(30)

# Dates d'acquisition, nombre d'oeuvres
requete = """
    SELECT o.date_acquisition,
           COUNT(o.id_oeuvre) AS Total_Oeuvres,
           ROUND(ROUND(COUNT(o.id_oeuvre),2)*100/(SELECT ROUND(COUNT(o.id_oeuvre),2) FROM oeuvre o),2) || ' %' AS Part_Oeuvres,
           COUNT(DISTINCT o.Id_Museo) AS Nb_Musees
           
    FROM oeuvre o,
    	musee m
 	
    WHERE o.Id_Museo = m.Id_Museo
    AND date_acquisition != 'NaN'
    AND date_acquisition NOT LIKE '%acquisition%'

    GROUP BY date_acquisition
    ORDER BY total_oeuvres DESC;
"""
date_acquisition = pd.read_sql_query(requete, conn)
print(date_acquisition.shape)
date_acquisition.head()

# Poucentage d'oeuvres exposées par villes, total oeuvres et musées
# Filtre sur un artiste, modifier la sous requête SELECT ROUND et WHERE requête principale
requete = """
    SELECT m.ville,
           a.nom_artiste AS Nom_Artiste,
           COUNT(DISTINCT a.id_museo) AS Nb_Musees,
           COUNT(a.id_oeuvre) AS Total_Oeuvres,
           ROUND(ROUND(COUNT(a.id_oeuvre),2)*100/(SELECT ROUND(COUNT(a.id_oeuvre),2) 
                                                  FROM v_Oeuvres_Artistes a
                                                  WHERE a.nom_artiste LIKE '%RENOIR%'),2) || ' %' AS                                                   Part_Oeuvres
    FROM musee m,
         v_Oeuvres_Artistes a

    WHERE a.Id_Museo = m.Id_Museo
    AND a.nom_artiste LIKE '%RENOIR%'
    AND ville = 'Paris'
    GROUP BY ville, a.nom_artiste

    ORDER BY nb_musees DESC;
"""
Villes_Oeuvres_Artiste = pd.read_sql_query(requete, conn)
print(Villes_Oeuvres_Artiste.shape)
Villes_Oeuvres_Artiste.head(30)

# Total oeuvres par domaine et musées
requete = """
    SELECT oc.Nom_Categorie,
           COUNT(oc.Id_oeuvre) AS Total_Oeuvres,
           COUNT(DISTINCT oc.id_museo) AS Nb_Musees
    FROM v_Oeuvres_Categories oc
 	
    GROUP BY Nom_Categorie
    ORDER BY nb_musees DESC;
"""
Categories_Oeuvres = pd.read_sql_query(requete, conn)
print(Categories_Oeuvres.shape)
Categories_Oeuvres.head()

# Total oeuvres par domaine et musées
# Par domaine et villes
# Recherche par ville et/ou domaine

requete = """
    SELECT m.ville,
           oc.Nom_Categorie,
           COUNT(oc.id_oeuvre) AS Total_Oeuvres,
           COUNT(DISTINCT oc.id_museo) AS Nb_Musees
    FROM v_Oeuvres_Categories oc
 	
    JOIN musee m
        ON oc.id_museo = m.id_museo

    WHERE ville LIKE '%Nantes%' AND Nom_categorie LIKE '%PEINTURE%'

    GROUP BY Nom_Categorie, m.ville;
"""
Categories_villes = pd.read_sql_query(requete, conn)
print(Categories_villes.shape)
Categories_villes.head()

# Total oeuvres par domaine et musées
# Par domaine et villes
# Recherche par ville et/ou domaine

requete = """
            SELECT oc.nom_categorie,
            COUNT(oc.id_oeuvre) as Nombre_Oeuvres,
            COUNT(distinct oc.id_museo) as Nombre_Musees,
            m.geo_x,
            m.geo_y

        FROM v_Oeuvres_Categories oc

        JOIN musee m
                ON oc.id_museo = m.id_museo
        
            WHERE geo_x != 'Nan' AND geo_y != 'Nan' 
            AND oc.nom_categorie LIKE '%%'
            GROUP BY oc.nom_categorie, oc.nom_categorie, m.geo_x, m.geo_y;
"""
Categories_geo = pd.read_sql_query(requete, conn)
print(Categories_geo.shape)
Categories_geo.head()

# Recherche doublons Domaine

requete = """
    SELECT c.domaine,
           COUNT(c.domaine) AS Nb_Domaine
    FROM categorie c
    GROUP BY c.domaine
    ORDER BY Nb_Domaine DESC;
"""
Categories_check = pd.read_sql_query(requete, conn)
print(Categories_check.shape)
Categories_check.head()

# Recherche doublons Oeuvre_Artiste

requete = """
    SELECT CONCAT(id_oeuvre, ' ', id_artiste) AS ids,
           COUNT(CONCAT(id_oeuvre, ' ', id_artiste)) AS Nb_ids
    FROM oeuvre_artiste
    GROUP BY ids
    ORDER BY Nb_ids DESC;
"""
OA_check = pd.read_sql_query(requete, conn)
print(OA_check.shape)
OA_check.head()

## SUPPRIMER VUE

# Effacer une vue
#functions_PgSql.executer_requete("DROP VIEW IF EXISTS v_Oeuvres_Categories;", conn)
